api/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.3.4; python_full_version >= '3.6.2'
azure-core==1.20.1
azure-storage-blob==12.9.0
azure-storage-queue==12.1.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7; python_version >= '3'
cryptography==36.0.0; python_version >= '3.6'
fastapi==0.70.0
idna==3.3; python_version >= '3.5'
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1; python_version >= '3.6'
pycparser==2.21
pydantic==1.8.2; python_full_version >= '3.6.1'
python-environ==0.4.54
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.16.0; python_version >= '3.6'
typing-extensions==4.0.0; python_version >= '3.6'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
